refactor: button types
